third_party/blink/web_tests/external/wpt/mathml/relations/html5-tree/tabindex-focus-001-expected.txt

This is a testharness.js-based test.
[FAIL] invalid tabindex attribute does not make the element focusable
  assert_equals: <a href=#> should be focusable expected Element node <a href="#" data-focusable="true"></a> but got Element node <body><math data-focusable="false"></math>\n<math href="#"...
Harness: the test ran to completion.
